Config Message

+++ Thermometer/Models/Config.cs
@@ -14,7 +14,7 @@ namespace Thermometer.Models
     public List<string> Emails { get; set; }
     public List<string> PhoneNumbers { get; set; }
 
-    public int DataRefreshRateInSec { get; set; }
+    public double DataRefreshRateInSec { get; set; }
 
     private Config()
     {
